Allow defining url domains to be used with specific credentials.
[ssu] Allow defining url domain to be used with specific credentials. Fixes JB#32346

Previously store-auth-repos could be only defined in repos.ini and each
new feature would need to add new line to that variable. This patch
allows repos.ini to define url domain,credential pairs that can be used
to say which credential is used when certain url domain is used.

This removes the need to touch repos.ini when bringing in new feature.

Signed-off-by: Marko Saukko <marko.saukko@jolla.com>