Suggest developer-mode to rnd images by default.
[ks] Suggest developer-mode to rnd images by default. Contributes to JB#39097

Other option would have been to create separate img-features-rnd= for
the rnd side, however that seems a bit overkill as in general the
features should be the same on release and rnd to make sure that we have
same functionality. The installation of developermode to the rnd images
is quite special case.

Signed-off-by: Marko Saukko <marko.saukko@jolla.com>