[ssu] Bump version to 1.0.0. Contributes to MER#1992
Earlier pkgconfig commit changed the .so version from so.1 to so.0.
Let's upgrade the version now to keep binary compatibility.

Not necessarily to be seen that API/ABI is set in stone, but as a reminder
there is at sailfish side some aliendalvik usage that needs to keep
working, namely SsuDeviceInfo::displayName().