[ssu] Don't close ssu daemon immediately on quit request. JB#62821
The client which is about to quit and suggest daemon to do the same
cannot know if there are other clients. Furthermore, if the client
is the ssu command line tool, it cannot know if it's being used in a
script where there are more invokations following.

So instead of quitting immediately, just fast-forward the normal autoquit
timer to 5 seconds if not below that already.